requirements.txt
fastapi
uvicorn
qdrant-client
langchain
llama-index
pydantic
python-dotenv
tiktoken
streamlit
requests
groq
openai>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
PyPDF2>=3.0.0
chardet>=5.0.0
openpyxl>=3.1.0
xlrd>=2.0.0
